requirements.txt
google-genai
gradio
faiss-cpu
pypdf
python-docx
python-dotenv
numpy
